+++ database/migrations/2023_05_24_003805_create_products_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('nama_pemilik');
             $table->string('nama_kos');
             $table->string('lokasi_kos');
-            $table->integer('harga_kos');
+            $table->double('harga_kos');
             $table->string('spesifikasi_kamar');
             $table->string('fasilitas_kamar');
             $table->string('fasilitas_umum');
